server_py/requirements.txt
fastapi
uvicorn
pandas
openpyxl
google-generativeai
python-dotenv
pydantic
supabase
python-multipart
